Move 'DiscreteDistribution' to utilities
(cherry picked from commit 8a49e298b7596184c389a27b5b7dc4522201ab0b)
